pyproject.toml
[tool.poetry]
name = "telegramchatbot"
version = "0.1.0"
description = ""
authors = ["Artem Sudoma <artem.sudoma@appfollow.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^21.3"
pydantic-settings = "^2.3.3"
uvloop = "^0.19.0"
openai = "^1.35.3"
motor = "^3.4.0"
sentry-sdk = "^2.7.1"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
flake8 = "^7.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ["py312"]

[tool.isort]
profile = "black"
